+++ example/android/app/src/main/kotlin/org/cloudplayplus/onscreen_gamepad_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.onscreen_gamepad_example
+package org.cloudplayplus.onscreen_gamepad_example
 
 import io.flutter.embedding.android.FlutterActivity